Move Target/Shortcut to new store.ts
