+++ src/jam.rs
@@ -1,4 +1,4 @@
-use std::{collections::HashMap, slice::Iter};
+use std::collections::HashMap;
 
 use anyhow::{anyhow, bail};
 use daggy::{Dag, NodeIndex, Walker};
@@ -8,29 +8,10 @@ use thiserror::Error;
 
 use crate::{
     config::{DesugaredConfig, DesugaredTargetCfg, Options},
-    executor::{ExecuteKind, Executor},
+    executor::Executor,
     reconciler::reconcile,
+    store::{Shortcut, Target},
 };
-
-struct Target<'a> {
-    name: &'a str,
-    shortcut: Shortcut,
-    help: &'a str,
-    cmd: Option<&'a str>,
-    execute_kind: ExecuteKind,
-}
-
-impl<'a> From<&'a DesugaredTargetCfg> for Target<'a> {
-    fn from(cfg: &'a DesugaredTargetCfg) -> Self {
-        return Target {
-            name: &cfg.name,
-            shortcut: Shortcut::from_shortcut_str(&cfg.shortcut_str),
-            help: &cfg.help,
-            cmd: cfg.cmd.as_deref(),
-            execute_kind: cfg.execute_kind,
-        };
-    }
-}
 
 type IdxT = u32;
 pub type NodeIdx = NodeIndex<IdxT>;
@@ -51,95 +32,6 @@ pub struct Jam<'a> {
     dag: Dag<Target<'a>, IdxT>,
     shortcuts: ShortcutTrie,
     logger: slog::Logger,
-}
-
-#[derive(Eq, Debug, PartialOrd, Ord)]
-pub struct Shortcut(pub Vec<char>);
-
-impl Shortcut {
-    pub fn empty() -> Shortcut {
-        Shortcut(vec![])
-    }
-
-    fn from_shortcut_str(shortcut_str: &str) -> Shortcut {
-        Shortcut(
-            shortcut_str
-                .split('-')
-                .map(|s| s.chars().next().unwrap())
-                .collect(),
-        )
-    }
-
-    // NOTE: This is not very efficient. It basically clones the shortcut
-    // then appends a key.  A more efficient approach is to implement
-    // ShortcutView or ShortcutChain or idek, but something that just takes
-    // a &Shortcut and &String and pretends like its the concatenation of
-    // the two without actually doing the concatenation.
-    pub fn append(&self, key: &char) -> Shortcut {
-        let mut new_vec = self.0.clone();
-        new_vec.push(*key);
-        Shortcut(new_vec)
-    }
-
-    // NOTE: The same thing as described for append() above is worth considering
-    // here as well.
-    pub fn prepend(&self, key: &char) -> Shortcut {
-        let mut new_vec = vec![*key];
-        new_vec.extend_from_slice(&self.0);
-        Shortcut(new_vec)
-    }
-
-    // NOTE: Same thing as described for append() above is worth considering here.
-    pub fn tail(&self) -> (Shortcut, Option<char>) {
-        let mut new_vec = self.0.clone();
-        let key = new_vec.pop();
-        (Shortcut(new_vec), key)
-    }
-
-    pub fn len(&self) -> usize {
-        self.0.len()
-    }
-
-    pub fn pop(&mut self) {
-        self.0.pop();
-    }
-
-    pub fn iter(&self) -> Iter<'_, char> {
-        self.0.iter()
-    }
-}
-
-impl From<&Self> for Shortcut {
-    fn from(value: &Self) -> Self {
-        Shortcut(value.0.clone())
-    }
-}
-
-impl From<Vec<&char>> for Shortcut {
-    fn from(value: Vec<&char>) -> Self {
-        Shortcut(value.iter().map(|c| **c).collect())
-    }
-}
-
-impl PartialEq for Shortcut {
-    fn eq(&self, other: &Self) -> bool {
-        self.0
-            .iter()
-            .zip(other.0.iter())
-            .fold(true, |acc, (s, o)| acc && s == o)
-    }
-}
-
-impl std::fmt::Display for Shortcut {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "{}", self.0.iter().intersperse(&'-').collect::<String>())
-    }
-}
-
-impl Clone for Shortcut {
-    fn clone(&self) -> Self {
-        Shortcut(self.0.to_vec())
-    }
 }
 
 #[derive(Error, Debug, PartialEq)]
@@ -955,7 +847,7 @@ mod tests {
     // use some mocking, however, the way ExecuteKind is right now
     // makes that annoying.
     mod execution {
-        use crate::{reconciler::Strategy, testutils::tmp::*};
+        use crate::{executor::ExecuteKind, reconciler::Strategy, testutils::tmp::*};
 
         use super::*;
